+++ src/main/java/app/service/IStatusService.java
@@ -6,6 +6,6 @@ import java.util.List;
 
 public interface IStatusService extends IOperations<TaskStatus> {
 
-    List<TaskStatus> findByPerformerId(int id);
+    List<TaskStatus> findByPerformerId(long id);
 
 }
